chore: move files to slideshow folder
