+++ src/pages/view/slideshow/SlideshowControlbar.jsx
@@ -15,8 +15,8 @@ import {
 import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React, { useState, useRef } from 'react'
-import PauseIcon from './slideshow/PauseIcon.jsx'
-import PlayIcon from './slideshow/PlayIcon.jsx'
+import PauseIcon from './PauseIcon.jsx'
+import PlayIcon from './PlayIcon.jsx'
 import { SlideshowFiltersInfo } from './SlideshowFiltersInfo.jsx'
 import styles from './styles/SlideshowControlbar.module.css'
 import { useSlideshowAutoplay, timingOptions } from './useSlideshowAutoplay.js'
