+++ src/pages/view/slideshow/useSlideshowAutoplay.js
@@ -4,7 +4,7 @@ import { useState, useEffect, useRef } from 'react'
 import {
     apiGetUserDataStoreValue,
     apiPostUserDataStoreValue,
-} from '../../api/userDataStore.js'
+} from '../../../api/userDataStore.js'
 
 const FIVE_SECONDS = '5'
 const TEN_SECONDS = '10'
